requirements.txt
duckdb
minio
python-dotenv
requests
pandas
fastparquet
dbt-core
dbt-duckdb
google-api-python-client
sqlalchemy
duckdb-engine